inputの値を親で変更できるように
